fix(tests): Properly set default route, fixed broken tests
By default `happy-dom` will have the URL `about:blank`. In the tests it
seems we were assuming we would immediately end up on an index page when
in fact we were not due to this behavior of `happy-dom`.

This change properly sets the default URL to `http://localhost:3000`
which is the same as Vitest sets for JSDOM.

There were two tests that started failing for this change:

1. "defaults to index page"
This test previously was ending up on the "slug" page, since the URL
was matching "about:blank" to this. This snapshot was updated since
it now correctly goes to the index page.

2. "allows pushing to other pages"
The `useRoute()` in this test can no longer be called sync. We need
to wait for the reactive route variable to update.