feat: register all auto-imports for mocking
It is not possible to mock auto-imports if they are registed with the
`imports:extend` hook. This is because nuxt-vitest also uses this hook to
register imports for mocking, but because there is no guaranteed call order of
the hook handlers, some auto-import registrations happen after nuxt-vitest's
hook handler.

This PR moves import registration to wait until Nuxt is ready, making all
auto-imports available for mocking.

There are some hints in the current implementation towards this being unwanted
behaviour, namely the `nuxtImportSources` whitelist introduced in 8806143.

I wanted to open this PR anyway to get feedback if there is some inherent
flaw in making all auto-imports available for mocking like this.