fix: fetch issues in JSDOM
Tests related to `fetch` and `$fetch` were failing due to problems
combining JSDOM and `node-fetch`: specifically `whatwg-url` and its
URLSearchParams implementation, and `node-fetch` incorrectly Proxying
URLSearchParams in its Headers implementation.

This change makes us use the window.fetch implementation if it exists
(as in the case of Happy-DOM). In JSDOM where it does not exist we
polyfill using whatwg-fetch which is standards compliant and (seems to)
work with whatwg-url.

Note: `globalThis.fetch` never worked in the code that was refactored.
Most likely we are in some sort of sandboxed environment inside Vitest
which makes us unable to access the Node globals like this.